requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
marshmallow==3.19.0
python-dotenv==1.0.0
pytest==7.3.1
gunicorn==20.1.0
